chore: finishing up save plugin and component
