+++ .github/workflows/test-on-pr-branch.yml
@@ -14,13 +14,14 @@ jobs:
           node-version: "16"
 
       - name: Install dependencies
-        run: npm ci
+        run: |
+          npm i @nx/nx-linux-x64-gnu
+          npm ci
 
       - name: Install playwright browsers
         run: npx playwright install --with-deps
 
-      - name: Build compas-open
-        working-directory: ./components/compas-open
+      - name: Build project
         run: npm run build
 
       - name: Run tests for compas-open
@@ -29,4 +30,12 @@ jobs:
 
       - name: Run tests for compas-open-plugin
         working-directory: ./packages/compas-open-plugin
+        run: npm run-script test
+
+      - name: Run tests for compas-save
+        working-directory: ./components/compas-save
+        run: npm run-script test
+
+      - name: Run tests for compas-save-plugin
+        working-directory: ./packages/compas-save-plugin
         run: npm run-script test
