+++ .github/workflows/test.yml
@@ -14,13 +14,14 @@ jobs:
           node-version: "16"
 
       - name: Install dependencies
-        run: npm ci
+        run: |
+          npm i @nx/nx-linux-x64-gnu
+          npm ci
 
       - name: Install playwright browsers
         run: npx playwright install --with-deps
 
-      - name: Build compas-open
-        working-directory: ./components/compas-open
+      - name: Build project
         run: npm run build
 
       - name: Run unit tests for compas-open
@@ -29,6 +30,14 @@ jobs:
 
       - name: Run unit tests for compas-open-plugin
         working-directory: ./packages/compas-open-plugin
+        run: npm run-script test:unit
+
+      - name: Run unit tests for compas-save
+        working-directory: ./components/compas-save
+        run: npm run-script test:unit
+
+      - name: Run unit tests for compas-save-plugin
+        working-directory: ./packages/compas-save-plugin
         run: npm run-script test:unit
   visual:
     runs-on: ubuntu-latest
@@ -42,7 +51,9 @@ jobs:
           node-version: "16"
 
       - name: Install dependencies
-        run: npm ci
+        run: |
+          npm i @nx/nx-linux-x64-gnu
+          npm ci
 
       - name: Install playwright browsers
         run: npx playwright install --with-deps
@@ -62,3 +73,19 @@ jobs:
         with:
           message: "ci: update screenshots"
           add: "components/compas-open/screenshots"
+
+      - name: Run visual test for compas-save
+        working-directory: ./components/compas-save
+        run: npm run-script test:visual
+
+      - name: Update screenshots for compas-save
+        if: failure()
+        working-directory: ./components/compas-save
+        run: npm run test:update
+
+      - name: Commit changes
+        if: failure()
+        uses: EndBug/add-and-commit@v9
+        with:
+          message: "ci: update screenshots"
+          add: "components/compas-save/screenshots"
